Prerelease refactoring
* Fixed RoleRegistrationService validation bug

* Added production db seeder

* Refactored app configuration

* Project clean up

+++ Backend/CoffeeScoutBackend.Api/Program.cs
@@ -1,31 +1,19 @@
 using CoffeeScoutBackend.Api;
-using CoffeeScoutBackend.Api.Config;
 using CoffeeScoutBackend.Api.DbSeeders;
 using CoffeeScoutBackend.Api.Extensions;
-using CoffeeScoutBackend.Api.Identity;
 using CoffeeScoutBackend.Api.Requests.Mappers;
 using CoffeeScoutBackend.Api.Responses.Mappers;
-using CoffeeScoutBackend.Bll;
+using CoffeeScoutBackend.Bll.Extensions;
 using CoffeeScoutBackend.Dal;
-using CoffeeScoutBackend.Dal.Config;
 using CoffeeScoutBackend.Dal.Entities;
-using FluentValidation;
-using MailerSendNetCore.Common.Extensions;
+using CoffeeScoutBackend.Dal.Extensions;
 using Microsoft.EntityFrameworkCore;
-using SharpGrip.FluentValidation.AutoValidation.Mvc.Extensions;
 
 var builder = WebApplication.CreateBuilder(args);
 
-builder.Services
-    .Configure<AdminSettings>(
-        builder.Configuration.GetSection(nameof(AdminSettings)))
-    .Configure<DatabaseSettings>(
-        builder.Configuration.GetSection(nameof(DatabaseSettings)))
-    .Configure<MailerSendSettings>(
-        builder.Configuration.GetSection(nameof(MailerSendSettings)));
-
 ResponseMapperConfiguration.Configure();
 RequestMapperConfiguration.Configure();
+builder.Services.AddApiSettings(builder.Configuration);
 
 builder.Services.AddEndpointsApiExplorer();
 builder.Services.AddSwagger();
@@ -35,47 +23,34 @@ builder.Services.AddProblemDetails();
 builder.Services.AddAuthentication();
 builder.Services.AddAuthorizationBuilder();
 
-var databaseSettings = builder.Configuration
-    .GetRequiredSection(nameof(DatabaseSettings))
-    .Get<DatabaseSettings>()!;
+builder.Services.AddTransient<IDbSeeder, ProdDbSeeder>();
+if (builder.Environment.IsDevelopment())
+{
+    builder.Services.AddTransient<IDbSeeder, TestDbSeeder>();
+    builder.Services.AddHttpLogging(_ => { });
+}
 
-builder.Services.AddTransient<IDbSeeder, TestDbSeeder>();
 builder.Services
     .AddIdentityServices()
     .AddBllServices()
-    .AddDalServices(databaseSettings);
+    .AddDalServices(builder.Configuration);
 
-builder.Services.AddMailerSendEmailClient(
-    builder.Configuration.GetSection(nameof(MailerSendSettings)));
-
-builder.Services.AddHttpContextAccessor();
-
-builder.Services.AddValidatorsFromAssemblyContaining<Program>();
-builder.Services.AddFluentValidationAutoValidation();
-
-builder.Services.AddHttpLogging(_ => { });
+builder.Services.AddApiInfrastructure(builder.Configuration);
+builder.Services.AddValidators();
 
 var app = builder.Build();
+using var scope = app.Services.CreateScope();
+var dbContext = scope.ServiceProvider.GetRequiredService<AppDbContext>();
 
 if (app.Environment.IsDevelopment())
 {
-    using var scope = app.Services.CreateScope();
-    var dbContext = scope.ServiceProvider.GetRequiredService<AppDbContext>();
     await dbContext.Database.EnsureDeletedAsync();
-    await dbContext.Database.MigrateAsync();
     app.UseSwagger();
     app.UseSwaggerUI();
-
     app.UseHttpLogging();
 }
 
-if (app.Environment.IsEnvironment("ApiTests"))
-{
-    using var scope = app.Services.CreateScope();
-    var dbContext = scope.ServiceProvider.GetRequiredService<AppDbContext>();
-    await dbContext.Database.EnsureDeletedAsync();
-    await dbContext.Database.MigrateAsync();
-}
+if (app.Environment.IsEnvironment("ApiTests")) await dbContext.Database.EnsureDeletedAsync();
 
 //app.UseHttpsRedirection();
 
@@ -90,6 +65,7 @@ app.AddExceptionHandlingMiddlewares();
 
 app.MapControllers();
 
+await dbContext.Database.MigrateAsync();
 var dbSeeder = app.Services.GetRequiredService<IDbSeeder>();
 await dbSeeder.SeedDbAsync();
